diary: refactoring all places doing link copying to use same and updated to ref structure
